Add post and like post

+++ KTUgram_backend/src/main/java/com/KTUgrammeriai/KTUgram_backend/post/PostRepository.java
@@ -2,6 +2,6 @@ package com.KTUgrammeriai.KTUgram_backend.post;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostRepository extends CrudRepository<Post, Integer> {
+public interface PostRepository extends CrudRepository<Post, Long> {
 
 }
